fix _findfirst handle type for win64 (#3035)
git-svn-id: https://svn.osgeo.org/gdal/branches/1.6@17270 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/port/cpl_vsil_win32.cpp
@@ -450,7 +450,7 @@ char **VSIWin32FilesystemHandler::ReadDir( const char *pszPath )
 
 {
     struct _finddata_t c_file;
-    long    hFile;
+    intptr_t hFile;
     char    *pszFileSpec, **papszDir = NULL;
 
     if (strlen(pszPath) == 0)
